Module Tezos_proxy

module Light : sig ... end

Code related to the light mode that is protocol-independent. See src/proto_*/lib_client/light.ml files for protocol-dependent code.

module Light_consensus : sig ... end

module Light_core : sig ... end

module Light_internal : sig ... end

module Light_logger : sig ... end

module Light_proto : sig ... end

module Logger : sig ... end

module Proxy : sig ... end

Code that is used both by protocol-dependent code and by other code. Note that we don't want this code in proxy_proto.ml because it's independent from the protocol and we neither want this code in proxy_getter.ml, because it would create a cyclic dependency between proxy_proto.ml and proxy_getter.ml

module Proxy_events : sig ... end

module Proxy_getter : sig ... end

module Proxy_proto : sig ... end

module Proxy_services : sig ... end

module RPC_client : sig ... end

module Registration : sig ... end

The module type of a proxy environment. Modules of this type should be prepared protocol-side and registered here to become available to the proxy facility.