Module Tezos_proxy.Proxy

Code that is used both by protocol-dependent code and by other code. Note that we don't want this code in proxy_proto.ml because it's independent from the protocol and we neither want this code in proxy_getter.ml, because it would create a cyclic dependency between proxy_proto.ml and proxy_getter.ml

module Local = Tezos_context_memory.Context

type proxy_getter_input = {

rpc_context : Tezos_rpc.RPC_context.simple;

chain : Tezos_shell_services.Block_services.chain;

block : Tezos_shell_services.Block_services.block;

}

type 'a update =

| Mutation

| Value of 'a

The result of setting a leaf. A mutation if done in place, otherwise a fresh value. We need this type because the proxy implementation returns a value whereas the light mode's implementation performs a mutation (because of Irmin under the hood).

module type TREE = sig ... end

module type CORE = sig ... end

Module used by implementations of Proxy_getter.M.