Module Tezos_proxy.Proxy_getter

module Local := Tezos_context_memory.Context

val raw_context_size : Tezos_shell_services.Block_services.raw_context -> int

The size of a tree, for logging

module StringMap = Tezos_stdlib.TzString.Map

module type REQUESTS_TREE = sig ... end

The point of this data structure is as follows:

module RequestsTree : REQUESTS_TREE

The point of this data structure is as follows:

module type M = sig ... end

type proxy_m = (module M)

module MakeProxy (X : Proxy_proto.PROTO_RPC) : M

Functor to obtain the implementation of M for the proxy mode (as opposed to the light mode implementation)

module Make (C : Proxy.CORE) (X : Proxy_proto.PROTO_RPC) : M

Functor to obtain a generic implementation. Used by the light mode

module Internal : sig ... end

Exposed for testing purpose only, you should not use it directly