Module Tezos_proxy.Proxy_services

type mode =

| Light_client of Light.sources
tezos-client --mode light is running

| Proxy_client
tezos-client --mode proxy is running

| Proxy_server of int option
tezos-proxy-server is running and the int option value is the value of argument --sym-block-caching-time

Whether using the light mode or the proxy mode (remember that the light mode is a different instance of the proxy mode (see srcs/lib_proxy/README_LIGHT.md for documentation) and whether tezos-client or tezos-proxy-server is running.

val build_directory : Tezos_client_base.Client_context.printer -> Tezos_rpc.RPC_context.json -> mode -> Registration.proxy_environment -> unit Tezos_rpc.RPC_directory.t

build_directory printer rpc_context env mode returns the directory of RPCs that is served locally by the client's light and proxy modes and by the proxy server. Parameters are:

printer is used for logging.

rpc_context is used to perform RPCs to distant endpoints.

mode specifies whether tezos-client (light or proxy mode) or tezos-proxy-server is running.

env is a protocol-specific module used to create the context passed when executing a RPC.