Module Tezos_proxy__Light_consensus

type input = {

printer : Tezos_client_base.Client_context.printer;

min_agreement : float;
The same value as Light.sources.min_agreement

chain : Tezos_shell_services.Block_services.chain;
The chain considered

block : Tezos_shell_services.Block_services.block;
The block considered

key : string list;
The key of the context for which data is being requested

mtree : Tezos_shell_services.Block_services.merkle_tree;
The tree received from the endpoint providing data. It is much smaller than tree.

tree : Tezos_context_memory.Context.tree;
The current data tree. The call to M.consensus will check that validating endpoints send data which agree with this tree.

}

A container of input data needed to process a consensus.

val min_agreeing_endpoints : float -> int -> int

min_agreeing_endpoints min_agreement nb_endpoints returns the minimum number of endpoints that must agree for Make.consensus to return true. The first parameter should be Light.sources.min_agreement while the second parameter should be the length of Light.sources.endpoints.

module Make : functor (Light_proto : Tezos_proxy.Light_proto.PROTO_RPCS) -> sig ... end

Given RPCs specific to the light mode, obtain the consensus algorithm