Module Tezos_proxy__Proxy_services

type mode =

| Light of Tezos_proxy.Light.sources

| Proxy

Whether using the light mode or the proxy mode (remember that the light mode is a different instance of the proxy mode (see srcs/lib_proxy/README_LIGHT.md for documentation).

val build_directory : Tezos_client_base.Client_context.printer -> Tezos_rpc.RPC_context.json -> mode -> Tezos_proxy.Registration.proxy_environment -> unit Tezos_rpc.RPC_directory.t

build_directory printer rpc_context mode env returns the directory of RPCs that is served locally by the proxy mode and the light mode. printer is used for logging. rpc_context is used to perform RPCs to distant endpoints. mode specifies whether to use the proxy mode or the light mode. env is a protocol-specific module used to create the context passed when executing a RPC.