Module Tezos_proxy__Registration

The module type of a proxy environment. Modules of this type should be prepared protocol-side and registered here to become available to the proxy facility.

module type Proxy_sig = sig ... end

type proxy_environment = (module Proxy_sig)

val register_proxy_context : proxy_environment -> unit

val get_registered_proxy : Tezos_client_base.Client_context.printer -> Tezos_rpc.RPC_context.#simple -> [< `Mode_light | `Mode_proxy ] -> Tezos_crypto.Protocol_hash.t option -> Tezos_shell_services.Block_services.chain -> Tezos_shell_services.Block_services.block -> proxy_environment Tezos_base__TzPervasives.tzresult Lwt.t

Returns a proxy environment for the given protocol (or the first one in the list of registered protocols if the Protocol_hash.t is None, see the Registration module).