Module Tezos_proxy.Light_internal

val key_to_string : string list -> string

Prints a key

val raw_context_to_irmin_tree : Tezos_shell_services.Block_services.raw_context -> Tezos_context_memory.Context.tree Lwt.t

Translates a raw_context, i.e. a tree sent by the raw_bytes RPC, to an Irmin tree that can be integrated in the light mode's store.

module Merkle : sig ... end

Module containing operations of Merkle proofs used by the light mode