Module Tezos_proxy.Proxy_getter

module Local := Tezos_context_memory.Context

val raw_context_size : Tezos_shell_services.Block_services.raw_context -> int

The size of a tree, for logging

module StringMap = Tezos_base__TzPervasives.TzString.Map

module type REQUESTS_TREE = sig ... end

module RequestsTree : REQUESTS_TREE

module type M = sig ... end

type proxy_m = (module M)

module MakeProxy : functor (X : Proxy_proto.PROTO_RPC) -> M

Functor to obtain the implementation of M for the proxy mode (as opposed to the light mode implementation)

module Make : functor (C : Proxy.CORE) -> functor (X : Proxy_proto.PROTO_RPC) -> M

Functor to obtain a generic implementation. Used by the light mode

module Internal : sig ... end

Exposed for testing purpose only, you should not use it directly